[release-branch.go1.21] cmd/internal/obj: hack FENCETSO to make it "fence rw, rw" instead
It is said that T-Head C910/C906 are early designs that don't support
fence.tso natively, so we would want to avoid emitting it because SBI
emulation is slow.

This is not currently used at present, but preemptively patch it so we
won't have to revisit this in the future.

The change is a hack, because the RevyOS Go fork is only intended to be
used with T-Head systems.

Closes: https://github.com/revyos/revyos/issues/22
Change-Id: I8107e3bd2d9c208cd0606f09b984ad59eb75010e